requirements.txt
absl-py==1.0.0
aiohttp==3.7.4.post0
alabaster==0.7.12
antlr4-python3-runtime==4.8
asttokens==2.0.5
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.2.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.12b0
boto3==1.20.46
botocore==1.23.46
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
coloredlogs==15.0.1
commonmark==0.9.1
Cython==0.29.27
dataclasses==0.6
datasets==1.18.2
debugpy==1.6.2
decorator==5.1.1
dill==0.3.4
discord-py-slash-command==3.0.3
discord.py==1.7.3
docformatter==1.4
docutils==0.15.2
editdistance==0.6.0
emoji==1.6.3
entrypoints==0.4
executing==0.8.2
fairscale==0.4.5
fairseq==0.10.0
filelock==3.4.2
fire==0.4.0
flake8==4.0.1
flake8-bugbear==22.1.11
fsspec==2022.1.0
gitdb==4.0.9
gitdb2==4.0.2
GitPython==3.1.26
google==3.0.0
google-api-core==2.8.2
google-auth==2.5.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.43.0
html2text==2020.1.16
html5lib==1.1
huggingface-hub==0.4.0
humanfriendly==10.0
hydra-core==1.1.1
idna==3.3
imagesize==1.3.0
importlib-metadata==4.2.0
iniconfig==1.1.1
iopath==0.1.9
ipykernel==6.15.1
ipython==8.0.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonlines==3.0.0
jupyter-client==7.3.4
jupyter-core==4.11.1
lxml==4.7.1
Markdown==3.3.2
markdown-it-py==0.5.8
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mock==4.0.3
multidict==6.0.2
multiprocess==0.70.12.2
mypy-extensions==0.4.3
myst-parser==0.12.10
nest-asyncio==1.5.5
nltk==3.6.7
numpy==1.21.0
oauthlib==3.1.1
omegaconf==2.1.1
packaging==21.3
pandas==1.4.0
parlai==1.6.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
portalocker==2.3.2
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-gfm==1.0.2
py-rouge==1.1
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.7
pytest==6.2.5
pytest-datadir==1.3.1
pytest-regressions==2.3.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
pyzmq==23.2.1
regex==2022.1.18
requests==2.27.1
requests-mock==1.9.3
requests-oauthlib==1.3.0
rich==11.1.0
rsa==4.8
s3transfer==0.5.0
sacrebleu==2.0.0
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.1.96
sh==1.14.2
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==2.2.2
sphinx-autodoc-typehints==1.10.3
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.1.4
subword-nmt==0.3.8
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.4.1
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.3
torch==1.11.0.dev20220126+cpu
torchtext==0.6.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.15.0
typing_extensions==4.0.1
Unidecode==1.3.2
untokenize==0.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
websocket-server==0.6.4
Werkzeug==2.0.2
xxhash==2.0.2
yarl==1.7.2
zipp==3.7.0
